srcs/api-gateway-app/requirements.txt
Flask==3.0.3
requests==2.32.4
pika==1.3.2
python-dotenv==1.0.1